LayoutTests/http/tests/dom/document-contentType-meta.html

<!DOCTYPE html>
<head>
 <title>document.contentType</title>
 <link rel="help" href="http://dom.spec.whatwg.org/#dom-document-contenttype">
 <meta http-equiv="content-type" content="text/xml">
 <script src="/js-test-resources/js-test-pre.js"></script>
</head>
<body>
 <script>
  shouldBeEqualToString('document.getElementsByTagName("META")[0].content', 'text/xml');
 </script>
 <script src="/js-test-resources/js-test-post.js"></script>
</body>
