Safari-611.1.10/LayoutTests/http/tests/dom/document-contentType-meta-expected.txt

PASS document.getElementsByTagName("META")[0].content is "text/xml"
PASS successfullyParsed is true

TEST COMPLETE
